requirements.txt
# Web Frameworks & UI
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.26.0

# HTTP & API Connectivity
requests>=2.31.0
httpx>=0.25.0

# Data Validation & Modeling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Analytics & Dashboard
pandas>=2.0.0

# Future Architecture (AI, RAG, & Web Scraping)
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
beautifulsoup4>=4.12.0
redis>=5.0.0
groq>=0.5.0
ollama>=0.1.0
python-dotenv>=1.0.0
pypdf>=5.0.0
sentence-transformers>=3.0.0
